make progress dialogs step count 64 bit
DirectoryMergeWindow::Data::fastFileComparison already tries to set the
step count with a 64 bit value. Make it an actual 64 bit value to avoid
integer overflow. While were at use const to guard against attempts to
change the inbound parameter.

+++ test/fakeprogressproxy.cpp
@@ -53,7 +53,7 @@ void ProgressProxy::step( bool bRedrawUpdate )
   (void)bRedrawUpdate;
 }
 
-void ProgressProxy::setMaxNofSteps( int dMaxNofSteps )
+void ProgressProxy::setMaxNofSteps( qint64 dMaxNofSteps )
 {
   /* Suppress warning about unused parameters */
   (void)dMaxNofSteps;
